Still very much a work in progress, but started on a crazy idea that supports mapping of classes/structures which in turns permits automatic tree generation and once in tree form it can then be serialised using json (or others once they're implemented).
The class previously called entity is now called tree since that
better represents what it is. The entity class is now something
to be derived from that supports mapped members.

Implemented vmaps which store a reference to the mapped class
member so that the value can be read from or written to.