Refactored vref so that objects and arrays are also handled as vrefs instead of vobj/varr. This re-arrangement should allow (although as yet untested) maps of vectors and vice verse.
Since the old tree structure is useful but no longer central to the encoding/decoding process it was possible to
implement a newer one, hopefully even more useful. It is temporarily called tree2 in this branch and supports
cool initialisation almost like dynamic objects in C#, should be more efficient than the old tree and uses
a similar method to pugixml attributes with as_*(defaultValue) calls for leaf nodes in the tree
(e.g., .as_string(), .as_long(42)). Tree encoding has been implemented, but decoding is to come.