An attempt to support encoding of const entities. Tidied up vref to put the type specific implementations in separate files. Added unit tests for the const side of things.
This results in almost identical ent_describe functions for the const/non-const implementations.

Warning: if you force the use of a const vref and attempt to decode it will silently do nothing. Perhaps
it should show an exception - but we don't usually use vref directly and the higher level ent::decode()
functions will assert at compile time if trying to decode a const thing.